(packaging) Update packaging for packaging repo PR 237
In https://github.com/puppetlabs/packaging/pull/237, we change how data is
stored/accessed in the packaging repo. This affects any workflows which use
these internal variables, and requires we update everything in tandem. An API
was never specified for the packaging repo, but if one was, PR 237 would
certainly constitute a break. However, after PR 237, we can move forward with
the work of actually defining said API.

This commit updates puppetdb's packaging internals to use the new entry points
in the packaging repo. The most invasive change is to the template method. The
erb functionality changed and became substantially more robust in PR 237 - as
such, we modify the template task in puppetdb to use the new method. However,
to DRY up some of the code, all of the calls to the templating method are
removed to one iteration over a hash of template => target pairs.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>