(packaging) Update variables in erbs to Pkg::Config
Some of the variables inside of the templates that puppetdb uses are defined
within the packaging repo. With the new packaging system, we've stopped
exporting all of the build configuration environment as global rake instance
variables, which means these configs cannot access them in the same way. In
other projects, we've solved this problem by providing the binding of the
Pkg::Config object when we erb - however, its not so simple with puppetdb. With
puppetdb, the templates use a mix of the variables from both puppetdb's
automation and the packaging repo's, and we cannot have both be accessible as
top-level @ instance variables without doing separate erb calls with different
bindings. Instead, we just keep the default top-level binding giving us access
to the puppetdb @ instance variables, since they're referenced much more than
those provided by the packaging repo. With the remaining @ instance variables
from the packaging repo, we fully qualify them as they exist in the binding.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>